+++ Otp/src/main/java/com/indusnet/repository/IUserModelRepository.java
@@ -1,11 +1,9 @@
 package com.indusnet.repository;
 
+import java.math.BigInteger;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 import com.indusnet.model.UserModel;
 
 @Repository
-public interface IUserModelRepository extends JpaRepository<UserModel, Integer> {
-
-}
+public interface IUserModelRepository extends JpaRepository<UserModel, BigInteger> {}
